Gave up on separate modules and order to make it work. There's much ordering and CR and documentation to do, but the main script works at least.Yay :)
