+++ research_tools/ObjectsCounter.py
