--- research_tools/PrototypicalObjectVerbMatcher/__init__.py
@@ -1,4 +1,0 @@
-import matcher
-import COCA
-import ObjectsCounter
-import synonyms
