--- research_tools/WordAbstractionEvaluator/DAL_AbstractionDB/__init__.py
@@ -1,1 +1,0 @@
-__author__ = 'Shay'
