--- research_tools/WordAbstractionEvaluator/Output.txt
@@ -1,59 +1,0 @@
-hot wet affectionate brief 
-loving tense public amicable 
-affectionate adoring Royal tense 
-comforting funny pleasant mixed 
-friendly split strong antisocial 
-friendly formal traditional usual 
-ardent public social financial 
-lively public current big 
-friendly big sly quick 
-lovable vivid old long 
-affectionate polite quiet mild 
-good golden wonderful vivid 
-gratifying political passionate ultimate 
-gentle big sugary strange 
-cute little beautiful honeyed 
-pleasing classical loud choral 
-pleasing old popular sad 
-lovable young sick small 
-satisfying fair bigger new 
-melodic male deep inner 
-pleasant vivid old long 
-gratifying Christian local Muslim 
-pleasant upset great electoral 
-intense terrible recurrent great 
-profound feigned archetypal cold 
-serious enormous powerful financial 
-profound tacit broad collective
-profound collective small audible 
-strong bottomless human real 
-profound private random happy 
-intense future human martial 
-low female human male 
-mysterious guarded soviet open 
-abstruse low economic cognitive 
-recondite public large free 
-gallus broad classic black
-wicked light Jewish ethnic 
-concealed little soviet open 
-evil intimate romantic intrusive 
-evil sole general main 
-pessimistic clear personal comic 
-threatening close fresh curious
-sinister good heroic past 
-vicious stone guardian Christmas 
-gloomy good positive public 
-evil armed social economic 
-sinister practical ancient powerful 
-difficult solid simple joint 
-dispassionate brief close firm 
-dispassionate solid good trustworthy 
-strong high political cold 
-strong firm low single 
-strong flavored organic exotic 
-intemperate young social oldest 
-tough good pure random 
-negative raw bad awful 
-convincing scant recent criminal 
-difficult insufficient wasted tough 
-difficult tough extra diligent
