--- research_tools/WordAbstractionEvaluator/__init__.py
@@ -1,1 +1,0 @@
-import WordAbstractionEvaluator
