+++ research_tools/synonyms.py
@@ -6,4 +6,4 @@ def get_synonyms_for(verb):
         return ["testsynonym"]
 
     synonym_sets = nltk.wordnet.wordnet.synsets(verb)
-    return [w.lemmas()[0].name() for w in synonym_sets].remove(verb)
+    return [w.lemmas()[0].name() for w in synonym_sets]
